requirements.txt
streamlit==1.33.0
google-generativeai==0.5.4
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.2.0
filetype==1.2.0
numpy==1.26.4
torch==2.1.2
transformers==4.30.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
safetensors==0.4.2
faiss-cpu==1.7.4
langchain==0.1.16
duckduckgo-search==5.3.1b1
gTTS==2.5.1